A major change to dimensions, units, and quantities
Review dimensions to determine if prime number or tuple encoding is better and consider a registry to store additional information or a faster compatibility matrix
Fixes #67

Create a fix to differentiate and correctly handle pound mass and pound force in the qnty library
Fixes #73

Revise how the unit "inch" is handled with "in" being a reserved keyword in Python. There is some inconsistency with specifying units as "inch" and being printed as "in".
Fixes #61